requirements.txt
hydra-core==1.3.2
haystack-ai==2.13.0
python-dotenv==1.1.0
pydantic==2.11.3
opensearch-haystack==4.0.0